fix: align README, examples, and build with actual codebase state
README claimed 7 nonexistent files (waker.mojo, coroutine.mojo,
executor.mojo, pending.mojo, time/, darwin/, windows/), cross-platform
support that isn't implemented, and a PendingOp type that doesn't exist.
Rewritten to reflect reality: stackful coroutines, ctypes/ bridge,
Linux-only with planned macOS/Windows.

Examples hardcoded x86_64 imports and used the deprecated
MutExternalOrigin alias. Switched to the arch-dispatched facade with
named syscall constants (__NR_read, __NR_write, __NR_close). Fixed
coro_echo which didn't compile since @explicit_destroy was added.

Build migrated from deprecated `mojo package` to `mojo precompile`
(.mojopkg → .mojoc). All compiler warnings resolved: implicit Int→UInt
conversions, lowercase docstring, unreachable dead code, unused
assignment. Added examples/ to the deprecated-origin check script.

+++ boucle/_sys/linux/io_uring/types.mojo
@@ -516,7 +516,7 @@ struct IoUringCqeFlags(TrivialRegisterPassable, Defaultable, Boolable):
         Returns:
             `self >> rhs`.
         """
-        return self.value >> rhs
+        return self.value >> UInt32(rhs)
 
 
 struct IoUringOp(TrivialRegisterPassable):
